requirements.txt
cffi          
cryptography  
getpass3    
oracledb    
pip         
pycparser 
python-dotenv 
setuptools    